refactor: 버튼 타입을 컴포넌트 파일로 이동
