+++ frontend/src/components/common/Button/Button.tsx
@@ -1,7 +1,17 @@
 import * as S from './Button.styled';
-import type { ButtonProps } from './ButtonProps';
+import type { PropsWithChildren, ButtonHTMLAttributes } from 'react';
 
-const Button = (props: ButtonProps) => {
+export type ButtonSize = 'small' | 'medium' | 'large';
+export type ButtonType = 'button' | 'submit' | 'reset';
+export type ButtonVariant = 'primary' | 'normal';
+
+export interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
+  size?: ButtonSize;
+  type?: ButtonType;
+  variant?: ButtonVariant;
+}
+
+const Button = (props: PropsWithChildren<ButtonProps>) => {
   const { children } = props;
 
   return <S.Button {...props}>{children}</S.Button>;
